dnn/io -> dnn/core
